+++ backend/src/test/java/com/cerebrus/cursoTest/CursoServiceImplTest.java
@@ -1,22 +1,21 @@
 package com.cerebrus.cursoTest;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.Mockito.never;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
-
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyString;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
+import static org.mockito.Mockito.never;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.security.access.AccessDeniedException;
 
@@ -624,9 +623,9 @@ class CursoServiceImplTest {
         when(usuarioService.findCurrentUser()).thenReturn(maestro);
         when(actividadAlumnoRepository.findByCursoID(10L)).thenReturn(actividades);
 
-        List<Integer> resultado = cursoService.obtenerNotaMediaPorActividadPorCursoId(10L);
+        List<Double> resultado = cursoService.obtenerNotaMediaPorActividadPorCursoId(10L);
 
-        assertThat(resultado).containsExactly(9, 7); // (8+10)/2 = 9, 7/1 = 7
+        assertThat(resultado).containsExactly(9.0, 7.0); // (8+10)/2 = 9, 7/1 = 7
     }
 
     // Test para verificar que obtenerNotaMediaPorActividadPorCursoId retorna lista vacía cuando no hay actividades
@@ -636,7 +635,7 @@ class CursoServiceImplTest {
         when(usuarioService.findCurrentUser()).thenReturn(maestro);
         when(actividadAlumnoRepository.findByCursoID(10L)).thenReturn(List.of());
 
-        List<Integer> resultado = cursoService.obtenerNotaMediaPorActividadPorCursoId(10L);
+        List<Double> resultado = cursoService.obtenerNotaMediaPorActividadPorCursoId(10L);
 
         assertThat(resultado).isEmpty();
     }
@@ -667,9 +666,9 @@ class CursoServiceImplTest {
         when(usuarioService.findCurrentUser()).thenReturn(maestro);
         when(actividadAlumnoRepository.findByCursoID(10L)).thenReturn(actividades);
 
-        List<Integer> resultado = cursoService.obtenerNotaMediaPorActividadPorCursoId(10L);
+        List<Double> resultado = cursoService.obtenerNotaMediaPorActividadPorCursoId(10L);
 
-        assertThat(resultado).containsExactly(0);
+        assertThat(resultado).containsExactly(0.0);
     }
 
     // -------------------------------------------------------
